% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amaTable.R
\name{amaTable}
\alias{amaTable}
\title{Blank and summarize growth data}
\usage{
amaTable(filename, everyHour = 2, blankName = "^blank.*",
  empty = "empty")
}
\arguments{
\item{filename}{A string that corresponds to the name fo the raw data file}

\item{everyHour}{How often should a data point be used, default is every other hour}

\item{blankName}{The regex expression used to find the blank wells}

\item{empty}{The regex expression used to find/remove empty wells}
}
\value{
A table of summarized and blanked readings. Writes this table to a csv file titled "analyzed_[your filename]"
}
\description{
Takes in the results from the machince reading, narrows down the time points and blanks all values.
}
\note{
Omits any NA data by simply removing that timepoint.
}
