% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare2all.r
\name{compare2all}
\alias{compare2all}
\title{Compare bacterial assays}
\usage{
compare2all(mydata, outGroup = "^PS392", alpha = 0.05, pval = F)
}
\arguments{
\item{mydata}{The output of amaTable. Required that headers are included for each well seperated by an underscore and in this format: Concentration_BacterialStrain}

\item{outGroup}{A regex expression that matches the non-wildtype strain. Default is PS392.}

\item{alpha}{The p-value below which results are no longer considered significant.}

\item{pval}{Raw p-values returned. By default is off.}
}
\value{
3 data frames that correspond to: Strain to strain, Wt to EtOH, Outgroup to EtOH. In that order.
}
\description{
Takes in the output of assay table then compares different strains at each timepoint to the same concentration using T-Tests.
}
\details{
Compares all concentrations to other strains and to the controls at each time point
}
