% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldStats.R
\name{ldStats}
\alias{ldStats}
\title{Take leaf disc information in long format and analyze}
\usage{
ldStats(db, check = F, logIt = F, twoVar = F)
}
\arguments{
\item{db}{Database of data in wide format}

\item{check}{False returns letter codes and assumption tests. True returns long format data, grouped data w std devs, the generated linear model, and raw statistical analysis.}

\item{logIt}{Log (base 10) the organized data before stats}

\item{twoVar}{Is there two independent variables? Default is FALSE.}
}
\value{
Returns letter code reults of ANOVA and also results of the statistical assumption tests
}
\description{
Takes in the results from the machince reading, narrows down the time points and blanks all values.
}
\note{
If ANOVA assumptions are failed, set logIt = T, then reset logIt and check the residuals of the linear model using check=T
}
